Looking into the problem of 4096 byte cookie sizes
So while cookies are nice, I'm really abusing them to store a list of filenames and timestamps and the max size is quickly exceeded after 70 or so files.  It looks like I'm going to have to stop fence-sitting between client and server and decide whether I'm going to more more towards the client or the server side.  Moving towards client-side, I will need to make the list of files accessible to the client, and use jStorage to remember what has been seen.  Moving towards server side, the file list will not need to be accessible to the client, but I will need to store information in a database.  This will make it easier to do more things with the data later, as I'll have all of it vs the clients.